Move BuilderState::createFilterOperations() to a separate file https://bugs.webkit.org/show_bug.cgi?id=246675
Reviewed by Antti Koivisto.

Make this function a stand-alone function and make used by the CSS style builder.
The plan is to use this function the 2D canvas for the filter API.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSFilterImageValue.cpp:
(WebCore::CSSFilterImageValue::createFilterOperations):
* Source/WebCore/style/ColorFromPrimitiveValue.cpp: Added.
(WebCore::colorFromPrimitiveValue):
(WebCore::colorFromPrimitiveValueWithResolvedCurrentColor):
* Source/WebCore/style/ColorFromPrimitiveValue.h: Added.
* Source/WebCore/style/FilterOperationsBuilder.cpp: Added.
(WebCore::filterOperationForType):
(WebCore::createFilterOperations):
* Source/WebCore/style/FilterOperationsBuilder.h: Added.
* Source/WebCore/style/StyleBuilderConverter.h:
(WebCore::Style::BuilderConverter::convertFilterOperations):
* Source/WebCore/style/StyleBuilderState.cpp:
(WebCore::Style::BuilderState::createFilterOperations):
(WebCore::Style::BuilderState::colorFromPrimitiveValue const):
(WebCore::Style::BuilderState::colorFromPrimitiveValueWithResolvedCurrentColor const):
(WebCore::Style::filterOperationForType): Deleted.
* Source/WebCore/style/StyleBuilderState.h:

Canonical link: https://commits.webkit.org/255706@main
